feat: 로그인 스타일 및 구성 개선, 유효성 검사 로직 추가
